CLN: move pyfibre_runner.py to main directory
