+++ pyfibre/addons/shg_pl_trans/tests/test_shg_pl_trans_analyser.py
@@ -4,7 +4,7 @@ from tempfile import TemporaryDirectory
 import numpy as np
 from pandas import DataFrame, Series
 
-from pyfibre.core.pyfibre_runner import PyFibreRunner
+from pyfibre.pyfibre_runner import PyFibreRunner
 from pyfibre.model.objects.segments import (
     FibreSegment, CellSegment
 )
